Move processTestFile to tests/helpers/comparison folder
