+++ packages/core/comparison/cli/checkPncMatching.ts
@@ -1,8 +1,8 @@
 import type PncComparisonResultDetail from "../types/PncComparisonResultDetail"
 
+import { parseComparisonFile } from "../../tests/helpers/comparison/processTestFile"
 import { isPhase1 } from "../lib/checkPhase"
 import comparePncMatching from "../lib/comparePncMatching"
-import { parseComparisonFile } from "../lib/processTestFile"
 import getStandingDataVersionByDate from "./getStandingDataVersionByDate"
 
 const checkPncMatching = async (
