+++ packages/core/comparison/cli/processFile.ts
@@ -1,10 +1,10 @@
 import type ComparisonResultDetail from "../types/ComparisonResultDetail"
 
+import { parseComparisonFile } from "../../tests/helpers/comparison/processTestFile"
 import { isPhase1, isPhase2, isPhase3 } from "../lib/checkPhase"
 import comparePhase1 from "../lib/comparePhase1"
 import comparePhase2 from "../lib/comparePhase2"
 import comparePhase3 from "../lib/comparePhase3"
-import { parseComparisonFile } from "../lib/processTestFile"
 import getStandingDataVersionByDate from "./getStandingDataVersionByDate"
 
 const processFile = async (
