+++ packages/core/comparison/lib/compareFile.ts
@@ -6,13 +6,13 @@ import getFileFromS3 from "@moj-bichard7/common/s3/getFileFromS3"
 import type ComparisonResult from "../types/ComparisonResult"
 import type ComparisonResultDetail from "../types/ComparisonResultDetail"
 
+import { parseComparisonFile } from "../../tests/helpers/comparison/processTestFile"
 import getStandingDataVersionByDate from "../cli/getStandingDataVersionByDate"
 import { isPhase1, isPhase2, isPhase3 } from "./checkPhase"
 import comparePhase1 from "./comparePhase1"
 import comparePhase2 from "./comparePhase2"
 import comparePhase3 from "./comparePhase3"
 import getDateFromComparisonFilePath from "./getDateFromComparisonFilePath"
-import { parseComparisonFile } from "./processTestFile"
 
 const failResult: ComparisonResultDetail = {
   auditLogEventsMatch: false,
