+++ packages/core/tsconfig.build.json
@@ -15,6 +15,7 @@
     "tests/helpers/comparison/parseIncomingMessage.ts",
     "tests/helpers/comparison/summariseMatching.ts",
     "tests/helpers/comparison/normalisePncOperations.ts",
+    "tests/helpers/comparison/processTestFile.ts",
     "tests/types/ComparisonFile.ts",
   ]
 }
